rename generator, improve docs
