+++ test/kafka_queue_buster_reader/fetcher-spec.ts
@@ -1,7 +1,7 @@
 import 'jest-extended';
 import { JobConfigParams } from '@terascope/job-components';
 import { WorkerTestHarness } from 'teraslice-test-harness';
-import { QueueBusterSliceRequest } from '../../asset/src/kafka_queue_buster_reader/interfaces.js';
+import { QueueBusterSliceRequest } from '../../asset/src/kafka_queue_buster_generator/interfaces.js';
 
 describe('kafka_queue_buster_reader fetcher', () => {
     let harness: WorkerTestHarness;
