+++ test/kafka_queue_buster_reader/schema-spec.ts
@@ -1,7 +1,7 @@
 import 'jest-extended';
 import { JobConfigParams } from '@terascope/job-components';
 import { WorkerTestHarness } from 'teraslice-test-harness';
-import { KafkaQueueBusterConfig } from '../../asset/src/kafka_queue_buster_reader/interfaces.js';
+import { KafkaQueueBusterConfig } from '../../asset/src/kafka_queue_buster_generator/interfaces.js';
 
 describe('kafka_queue_buster_reader schema', () => {
     let harness: WorkerTestHarness;
